+++ src/novafabric/serve/static/_astro/CapsuleInspector.B7UWFLzs.js
@@ -1,0 +1,1 @@
+import{t as e}from"./CapsuleInspector.BT4brWdn.js";export{e as default};