+++ src/novafabric/serve/static/_astro/CapsuleInspector.DF5FtRHF.js
@@ -1,1 +1,0 @@
-import"./jsx-runtime.u17CrQMm.js";import"./index.Bcm1Uarj.js";import"./clsx.B-dksMZM.js";import"./fixtures.q_89pNCY.js";import"./api.BweFz0av.js";import{C as f}from"./CapsuleInspector.hXWAwy7K.js";import"./runUtils.BhaFw2bI.js";export{f as default};