+++ src/novafabric/serve/static/_astro/DashboardApp.D5_AB1HI.js
@@ -1,1 +1,0 @@
-import{D as i}from"./DashboardApp.XXMdmHyw.js";import"./jsx-runtime.u17CrQMm.js";import"./index.Bcm1Uarj.js";import"./api.BweFz0av.js";export{i as default};