+++ src/novafabric/serve/static/_astro/index.0hVYzVNy.js
@@ -1,1 +1,0 @@
-import{r}from"./index.CUO2cybH.js";var a=r();export{a as r};